新增pcg树木和terrain的远景GI初步计算 优化编辑器的选择功能 新增prefab生成和编辑功能

+++ EngineAssets/Shaders/PunctualShadow/AlphaTest/PunctualShadowAlphaTest.vert
@@ -1,15 +1,13 @@
 #version 450
 #extension GL_GOOGLE_include_directive : require
-#extension GL_EXT_shader_16bit_storage : require
-#extension GL_EXT_shader_explicit_arithmetic_types : require
 
 #include "../../Common/ModelData.glsl"
 #include "../../Common/VansDrawSubmission.glsl"
 #include "../../Common/VertexDeformation.glsl"
 #include "../../Lights/LightsData.glsl"
 
-layout(location = 0) in f16vec4 position;
-layout(location = 1) in f16vec2 uv;
+layout(location = 0) in vec4 position;
+layout(location = 1) in vec2 uv;
 
 layout(location = 0) out vec2 fragUV;
