+++ EngineAssets/Shaders/PunctualShadow/PunctualShadow.vert
@@ -1,15 +1,13 @@
 #version 450
 #extension GL_GOOGLE_include_directive : require
-#extension GL_EXT_shader_16bit_storage : require
-#extension GL_EXT_shader_explicit_arithmetic_types : require
 
 #include "../Common/ModelData.glsl"
 #include "../Common/VansDrawSubmission.glsl"
 #include "../Common/VertexDeformation.glsl"
 #include "../Lights/LightsData.glsl"
 
-layout(location = 0) in f16vec4 position;
-layout(location = 1) in f16vec3 normal;
+layout(location = 0) in vec4 position;
+layout(location = 1) in vec3 normal;
 
 void main()
 {
