+++ EngineAssets/Shaders/Shadow/Shadow.vert
@@ -1,7 +1,5 @@
 #version 450
 #extension GL_GOOGLE_include_directive : require
-#extension GL_EXT_shader_16bit_storage : require
-#extension GL_EXT_shader_explicit_arithmetic_types : require
 
 #define LightCBBind 0
 #include "../Common/ModelData.glsl"
@@ -9,8 +7,8 @@
 #include "../Common/VertexDeformation.glsl"
 #include "../Lights/LightsData.glsl"
 
-layout(location = 0) in f16vec4 position;
-layout(location = 1) in f16vec3 normal;
+layout(location = 0) in vec4 position;
+layout(location = 1) in vec3 normal;
 
 layout(location = 0) out float shadowDepth;
